+++ src/command_modules/azure-cli-consumption/azure/cli/command_modules/consumption/custom.py
@@ -27,7 +27,7 @@ def cli_consumption_list_usage(client, billing_period_name=None, top=None, inclu
     if billing_period_name and top:
         return list(client.list_by_billing_period(billing_period_name, expand=expand, filter=filter_expression, top=top).advance_page())
     elif billing_period_name and not top:
-        return list(client.list_by_billing_period(expand=expand, filter=filter_expression))
+        return list(client.list_by_billing_period(billing_period_name, expand=expand, filter=filter_expression))
     elif not billing_period_name and top:
         return list(client.list(expand=expand, filter=filter_expression, top=top).advance_page())
     return client.list(expand=expand, filter=filter_expression)