+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/commands.py
@@ -12,7 +12,7 @@ from azure.cli.core.commands import CliCommandType
 from ._client_factory import (_auth_client_factory, _graph_client_factory,
                               _msi_user_identities_operations, _msi_operations_operations)
 
-from ._validators import process_msi_namespace
+from ._validators import process_msi_namespace, process_assignment_namespace
 
 
 def transform_definition_list(result):
@@ -130,9 +130,9 @@ def load_command_table(self, _):
         g.custom_command('update', 'update_role_definition')
 
     with self.command_group('role assignment') as g:
-        g.custom_command('delete', 'delete_role_assignments')
-        g.custom_command('list', 'list_role_assignments', table_transformer=transform_assignment_list)
-        g.custom_command('create', 'create_role_assignment')
+        g.custom_command('delete', 'delete_role_assignments', validator=process_assignment_namespace)
+        g.custom_command('list', 'list_role_assignments', validator=process_assignment_namespace, table_transformer=transform_assignment_list)
+        g.custom_command('create', 'create_role_assignment', validator=process_assignment_namespace)
         g.custom_command('list-changelogs', 'list_role_assignment_change_logs')
 
     with self.command_group('ad app', client_factory=get_graph_client_applications, resource_type=PROFILE_TYPE,