+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -5,6 +5,7 @@ Release History
 
 * [BREAKING CHANGE]: previously deprecated command group `network interface-endpoints` has been removed. Use `network private-endpoints` instead.
 * `network vnet subnet create/update`: Added `--nat-gateway` argument for attaching to a NAT gateway.
+* `dns zone import`: Fix issue where record names could not match a record type.
 
 2.4.0
 +++++